third_party/blink/web_tests/external/wpt/fetch/local-network-access/fetch.tentative.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>LNA Fetch tests: HTTPS and Public source </title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/support.sub.js"></script>
<script>
  "use strict";

  promise_test(t => {
    const source = { server: Server.HTTPS_PUBLIC };
    const sourceUrl =
        resolveUrl("resources/fetch-private.html", sourceResolveOptions(source));

    function checkResult(evt) {
      checkTestResult(evt.data, FetchTestResult.SUCCESS);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Public to private with permission');

  promise_test(t => {
    const source = { server: Server.HTTPS_PUBLIC };
    const sourceUrl = resolveUrl("resources/fetch-private-permission-denied.html",
        sourceResolveOptions(source));

    function checkResult(evt) {
      checkTestResult(evt.data, FetchTestResult.FAILURE);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Public to private with permission denied');
</script>
</body>
